[Refactor][16th,17th Day] To PostgreSQL
Needed to create a new Entity that would be connected in a relationsip
OneTOMany and bind that via Spring.

For this reason most of the Backend project was refactored to try
a different approach

Added Lombok to generated the setters & Getters and create the 
hypertext and other utilities necessary for the REST endpoints.


Migrated the create-product.sql code from MYSQL to PostgreSQL
Adapted the script for my project.

Besides learning the diff between the two RDBMS:
Learned how to use BIGSerial isntead of BIGINT
Understood the use of the 3 valued BIT.(e.g.: if the user provides no input or false or true)


PosgtreSQL
changed the driver and the necessary settings, including, defining the schema
For every variable had to change the collumn name has PostgreSQL did not allow for camelCase format.
With lombak we already have the hyperlinks for representation and we also manage to have the
total count of pages and the pagination system created for us.